+++ gallery_generator/scripts/__init__.py
@@ -1,0 +1,6 @@
+import sys
+
+
+def exit_failure(msg: str, code: int = -1):
+    print(msg, file=sys.stderr)
+    return sys.exit(code)
